ingestion/requirements.txt
kafka-python
Pillow
moviepy==1.0.3
imageio[ffmpeg]